textbox requires OnChange for validation
fixes #6, fixes #7
for textbox validation TextBox_Change must be used so cut/paste and backspace/delete are captured

I thought I worked with the latest file but now when raising this pull request I ran a VBAdiff and so other (minor) changes that all did not fix #6 and #7.  So I uploaded my file for you to review and decide/amend.

another userform added (and left) for quickly testing textbox events